refactor: cleanup JSDoc and encapsulate internal types
